added ability to purchase photos

+++ src/com/dk/controller/AccountController.java
@@ -19,14 +19,18 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.dk.model.CreditCard;
+import com.dk.model.Order;
 import com.dk.model.Picture;
 import com.dk.model.PictureData;
 import com.dk.model.PictureThumb;
 import com.dk.model.User;
+import com.dk.service.OrderService;
 import com.dk.service.PictureDataService;
 import com.dk.service.PictureService;
 import com.dk.service.PictureThumbService;
 import com.dk.service.UserService;
+import com.dk.utility.PaymentTool;
 import com.dk.utility.PictureTool;
 import com.dk.utility.SessionTool;
 
@@ -36,18 +40,26 @@ public class AccountController {
 	private PictureService pictureService;
 	private PictureDataService pictureDataService;
 	private PictureThumbService pictureThumbService;
+	private OrderService orderService;
 	private UserService userService;
 	@Autowired
 	private SessionTool sessionTool;
+	@Autowired
+	private PaymentTool paymentTool;
 
 	public AccountController() {
-		sessionTool = new SessionTool();
 	}
 
 	@Autowired(required = true)
 	@Qualifier(value = "userService")
 	public void setUserService(UserService userService) {
 		this.userService = userService;
+	}
+
+	@Autowired(required = true)
+	@Qualifier(value = "orderService")
+	public void setOrderService(OrderService orderService) {
+		this.orderService = orderService;
 	}
 
 	@Autowired(required = true)
@@ -68,25 +80,51 @@ public class AccountController {
 		this.pictureThumbService = pictureThumbService;
 	}
 
+	/**
+	 * 
+	 * Used to allow user access to the account details page. The user has to be
+	 * authenticated in order to gain access to this resource.
+	 * 
+	 * The accounts page consists of 3 things: account details, uploaded photos,
+	 * and purchased photos. These fields are populated here and passed into the
+	 * account view.
+	 * 
+	 * @param model
+	 *            current model
+	 * @return
+	 */
 	@RequestMapping(value = "/account")
 	public String listUsers(Model model) {
-		System.out.println("account");
+
+		// verify that the user is logged in
 		if (sessionTool.getCurrentUser() == null) {
 			model.addAttribute("user", new User());
 			model.addAttribute("target", "account");
 			return "login";
 		} else {
+			// populate required fields to display account information
 			User tmpUser = userService.getUser(sessionTool.getCurrentUser().getUsername());
 			model.addAttribute("user", tmpUser);
 			sessionTool.setCurrentUser(tmpUser);
 			model.addAttribute("uploadedPictures", pictureService.getByUsername(tmpUser));
+			model.addAttribute("orders", orderService.getByUsername(tmpUser));
 			return "account";
 		}
 	}
 
+	/**
+	 * 
+	 * Used to redirect the user to the upload page. The user must be
+	 * authenticated in order to access this resource.
+	 * 
+	 * @param model
+	 *            current model
+	 * @return
+	 */
 	@RequestMapping(value = "/upload")
 	public String upload(Model model) {
-		System.out.println("upload");
+
+		// verify that the user is logged in
 		if (sessionTool.getCurrentUser() == null) {
 			model.addAttribute("user", new User());
 			model.addAttribute("target", "upload");
@@ -96,32 +134,76 @@ public class AccountController {
 		}
 	}
 
+	/**
+	 * 
+	 * Used to direct the user to the purchase form. The user must be
+	 * authenticated in order to access this resource.
+	 * 
+	 * The id of the photo the user wishes to purchase must be passed in as a
+	 * request parameter.
+	 * 
+	 * @param model
+	 *            current model
+	 * @param id
+	 *            the id of the photo to be purchased
+	 * @return
+	 */
 	@RequestMapping(value = "/buy{id}")
-	public String buy(Model model, @RequestParam String id) {
-		System.out.println("buy; id=" + id);
+
+	// verify that the user is logged in
+	public String buy(Model model, @RequestParam int id) {
 		if (sessionTool.getCurrentUser() == null) {
 			model.addAttribute("user", new User());
 			model.addAttribute("target", "buy");
 			return "login";
 		} else {
+			model.addAttribute("picture", pictureService.getPicture(id));
 			return "buy";
 		}
 	}
-	
+
+	/**
+	 * 
+	 * Used to delete a picture from the gallery. The user attempting to delete
+	 * a photo must be logged in. A user can only delete photos they uploaded.
+	 * 
+	 * The photo data and the photo thumbnail must be deleted before the photo
+	 * in order to satisfy foreign key restraints in the database. After the
+	 * photo and all related data is deleted, the user is redirected back to the
+	 * account page.
+	 * 
+	 * @param model
+	 *            current model
+	 * @param id
+	 *            the id of the photo to be deleted
+	 * @return
+	 */
 	@RequestMapping(value = "/deletePicture{id}")
 	public String deletePicture(Model model, @RequestParam int id) {
+		if (sessionTool.getCurrentUser() == null) {
+			model.addAttribute("user", new User());
+			model.addAttribute("target", "account");
+			return "login";
+		}
+
 		Picture toDelete = pictureService.getPicture(id);
 		PictureData toDeleteData = pictureDataService.getPictureData(toDelete.getId());
 		PictureThumb toDeleteThumb = pictureThumbService.getPictureThumb(toDelete.getId());
-		
-		if(toDelete.getUsername().getUsername().equals(sessionTool.getCurrentUser().getUsername())) {
+
+		// verify that the user attempting to delete a picture is the same user
+		// that uploaded it
+		if (toDelete.getUsername().getUsername().equals(sessionTool.getCurrentUser().getUsername())) {
+			// delete all picture data in order of dependencies
 			pictureThumbService.deletePictureThumb(toDeleteThumb);
 			pictureDataService.deletePictureData(toDeleteData);
 			pictureService.deletePicture(toDelete);
 			model.addAttribute("successMsg", "Picture successfully deleted.");
 		} else {
+			// the user is not the one we expected
 			model.addAttribute("dangerMsg", "You lack permission to delete that picture.");
 		}
+
+		// redirect back to account
 		model.addAttribute("user", sessionTool.getCurrentUser());
 		model.addAttribute("uploadedPictures", pictureService.getByUsername(sessionTool.getCurrentUser()));
 		return "account";
@@ -134,6 +216,28 @@ public class AccountController {
 		return "redirect:/users";
 	}
 
+	/**
+	 * 
+	 * Used to upload a picture to the gallery. The user must be authenticated
+	 * in order to upload pictures.
+	 * 
+	 * After the user is verified to be authenticated, the picture itself is
+	 * verified to be a jpg. Then a thumbnail is created that will be displayed
+	 * in the gallery, and all the related data is stored into the database. The
+	 * user is then redirected back to the upload page wiht a success message.
+	 * 
+	 * If anything goes wrong, or the file fails acceptance criteria, the user
+	 * is redirected back to the upload page with a danger message.
+	 * 
+	 * @param picture
+	 *            the picture to be saved
+	 * @param model
+	 *            the current model
+	 * @param price
+	 *            the desired price of the photo
+	 * @return
+	 * @throws Exception
+	 */
 	@RequestMapping(value = "/upload", method = RequestMethod.POST)
 	public ModelAndView upload(@RequestParam("picture") MultipartFile picture, Model model,
 			@RequestParam("price") double price) throws Exception {
@@ -186,45 +290,110 @@ public class AccountController {
 		return mv;
 	}
 
+	/**
+	 * 
+	 * Used to log a user into the current session.
+	 * 
+	 * @param user
+	 *            login data
+	 * @param model
+	 *            current model
+	 * @param target
+	 *            the protected resource the user was attempting to access
+	 * @return
+	 */
 	@RequestMapping(value = "/login", method = RequestMethod.POST)
 	public String login(@ModelAttribute("user") User user, Model model, @RequestParam("target") String target) {
 
+		// verify the login was successful
 		User tmp = userService.getUser(user.getUsername());
 		if (tmp == null || tmp.getUsername() == null) {
 			model.addAttribute("dangerMsg", "Invalid username or password. Please try again.");
 			model.addAttribute("target", target);
 			return "login";
 		}
+
+		// add the user to the session and direct to the page they were trying
+		// to access
 		sessionTool.setCurrentUser(user);
 		model.addAttribute("user", tmp);
 		return "redirect:" + target;
 	}
-	
-	
+
+	/**
+	 * 
+	 * Used to register a new user.
+	 * 
+	 * @param user
+	 *            the user information to be persisted
+	 * @param model
+	 *            the current model
+	 * @param target
+	 *            the protected resource that the user was attempting to access
+	 * @return
+	 */
 	@RequestMapping(value = "/register", method = RequestMethod.POST)
 	public String register(@ModelAttribute("user") User user, Model model, @RequestParam("target") String target) {
 
+		// verify the registration won't break PK restraint
 		User tmp = userService.getUser(user.getUsername());
 		if (tmp != null) {
 			model.addAttribute("dangerMsg", "Username already exists.");
 			model.addAttribute("target", target);
 			return "login";
 		}
-		
+
+		// save the user and direct them to the page they were trying to access
 		userService.SaveOrUpdateUser(user);
-		
+
 		sessionTool.setCurrentUser(user);
 		model.addAttribute("user", tmp);
-		return target;
+		return "redirect:" + target;
 	}
-	
-	
-	
-	
+
+	/**
+	 * Used to log the current user out of the session.
+	 * 
+	 * @return redirect to the index page
+	 */
 	@RequestMapping(value = "/logout")
 	public String logout() {
 		sessionTool.setCurrentUser(null);
 		return "index";
+	}
+
+	/**
+	 * 
+	 * Used to verify the final purchase of a photo. The user must be
+	 * authenticated in order to access this resource.
+	 * 
+	 * @param model
+	 * @param cardNumber
+	 * @param expiryMonth
+	 * @param expiryYear
+	 * @param cvCode
+	 * @return
+	 */
+	@RequestMapping("/verifyPayment")
+	public String verifyPayment(Model model, @RequestParam("cardNumber") String cardNumber,
+			@RequestParam("expiryMonth") int expiryMonth, @RequestParam("expiryYear") int expiryYear,
+			@RequestParam("cvCode") int cvCode, @RequestParam("total") double total, @RequestParam("id") int id) {
+		
+		System.out.println("card=" + cardNumber);
+
+		CreditCard creditCard = new CreditCard(cardNumber, expiryMonth, expiryYear, cvCode);
+
+		if (paymentTool.processRequest(PaymentTool.TRANSACTION_TYPE.SALE, creditCard, total).equals("success")) {
+			Order order = new Order();
+			order.setPictureId(pictureService.getPicture(id));
+			order.setUsername(sessionTool.getCurrentUser());
+			orderService.saveOrUpdateOrder(order);
+			return "redirect:account";
+		} else {
+			model.addAttribute("dangerMsg", "Your transaction was not approved. Please try again.");
+		}
+
+		return "redirect:account";
 	}
 
 }
